Renamed styles file
